Forgot to push changes for renamed resources
